Refresh with shopping template as base
Shopping template is much more fleshed out than the main template so I went back and touched it up. Same exact functionality as before, just with little bug fixes here and there.

+++ Angular/src/app/dashboard/card.service.ts
@@ -49,9 +49,9 @@ export class CardService {
 
 	}
 
-	public deleteCard(id: string): Observable<string> {
+	public deleteCard(id?: string): Observable<string> {
 		const accessToken = localStorage.getItem('accessToken');
-		if (accessToken) {
+		if (accessToken && id) {
 			const headers = new HttpHeaders().append('Authorization', 'Bearer ' + accessToken).append('Accept', 'application/json');
 			return this.http.delete<string>(this.cardBase + 'card/' + id, { headers });
 		} else {
